geoseg_new/requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
mercantile>=1.2.0
rasterio>=1.2.0
supermercado>=0.3.0
geojson>=2.5.0
osmium>=3.0.0
shapely>=1.8.0
tqdm>=4.60.0
toml>=0.10.2
requests>=2.25.0
opencv-python>=4.5.0
scipy>=1.7.0

# Optional: For SAM annotation (trees, grass detection)
# Uncomment to enable --enable-sam feature

# For SAM 1 (original):
# torch>=2.0.0
# segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# For SAM 2 (recommended, better quality):
# torch>=2.0.0
# sam2 @ git+https://github.com/facebookresearch/segment-anything-2.git

